src/requirements.txt
aiohttp>=3.9.0
azure-ai-contentunderstanding>=1.0.0
azure-ai-documentintelligence>=1.0.0
azure-identity>=1.17.0
azure-mgmt-cognitiveservices>=13.0.0
pymupdf>=1.25.0
python-dotenv>=1.0.0
